[REF] rendering: avoid getter usage
This commit aims to make the drawing process more independent from the
getters. All the drawing of the grid is now done in helpers that
don't have the getters.

Task: 6432165
