+++ src/components/spreadsheet_print/spreadsheet_print_store.ts
@@ -182,6 +182,8 @@ export class SpreadsheetPrintStore extends SpreadsheetStore {
     });
     viewports.setSheetViewOffset(sheetId, firstColStart, firstRowStart);
 
+    const theme = this.getters.getSpreadsheetTheme();
+    const sheet = this.getters.getSheet(sheetId);
     return {
       sheetId,
       viewports,
@@ -191,6 +193,8 @@ export class SpreadsheetPrintStore extends SpreadsheetStore {
       activeCols: new Set(),
       activeRows: new Set(),
       activePosition: undefined,
+      theme,
+      backgroundColor: sheet.backgroundColor || theme.backgroundColor,
     };
   }
