+++ tests/test_helpers/renderer_helpers.ts
@@ -112,8 +112,23 @@ export class MockGridRenderingContext implements GridRenderingContext {
   }
 
   get hideGridLines(): boolean {
-    // Handled in the rendering context created by the `Dashboard` component in practice
-    return this.model.getters.isDashboard();
+    // Handled in the rendering context created by the `Dashboard`/`Grid` components in practice
+    return this.model.getters.isDashboard()
+      ? true
+      : !this.model.getters.getGridLinesVisibility(this.sheetId);
+  }
+
+  get dataBarVerticalPadding() {
+    // Handled in the rendering context created by the `Dashboard`/`Grid` components in practice
+    return this.model.getters.isDashboard() ? 2 : 0;
+  }
+
+  get theme() {
+    return this.model.getters.getSpreadsheetTheme();
+  }
+
+  get backgroundColor() {
+    return this.model.getters.getSheet(this.sheetId).backgroundColor || this.theme.backgroundColor;
   }
 }
